services/ai-service/requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
google-generativeai==0.8.5
python-dotenv==1.0.0
requests==2.32.5



